+++ Mcsg.Story.Api/Services/PostService.cs
@@ -2096,7 +2096,7 @@ public partial class PostService : IPostService
 
         var orders = await _context.StorySubPostAvailable.Where(p => p.PostId == postId).Select(p => p.Order).ToListAsync();
 
-        return orders.Count > 0 ? orders.Max() + 1 : 1;
+        return orders.Count > 0 ? (int)orders.Max() + 1 : 1;
     }
 
     /// <summary>
